refactor: 장바구니 아이템 컴포넌트를 재사용 가능하도록 type prop 생성
